# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: THREE NINE POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List all products
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/{id}/image:
    delete:
      operationId: deleteProductImage
      tags: [products]
      summary: Remove product image
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /stock-movements:
    get:
      operationId: getStockMovements
      tags: [inventory]
      summary: List stock movements
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
    post:
      operationId: createStockMovement
      tags: [inventory]
      summary: Record a stock movement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockMovementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"
  /recipes/{productId}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get recipe for a product
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertRecipe
      tags: [recipes]
      summary: Create or replace recipe for a product
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeInput"
      responses:
        "200":
          description: Upserted recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
  /recipes:
    get:
      operationId: getRecipes
      tags: [recipes]
      summary: List all recipes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recipes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        sellingPrice:
          type: number
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        hpp:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, code, categoryId, sellingPrice, status, hpp, createdAt, updatedAt]
    ProductInput:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        categoryId:
          type: integer
        description:
          type: string
        sellingPrice:
          type: number
        status:
          type: string
          enum: [active, inactive]
      required: [name, code, categoryId, sellingPrice]
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        categoryId:
          type: integer
        description:
          type: string
        sellingPrice:
          type: number
        status:
          type: string
          enum: [active, inactive]
...
    StockMovement:
      type: object
      properties:
        id:
          type: integer
        ingredientId:
          type: integer
        ingredientName:
          type: string
        unit:
          type: string
        movementType:
          type: string
          enum: [in, out, adjustment, opname]
        quantity:
          type: number
        previousStock:
          type: number
        newStock:
          type: number
        notes:
          type: ["string", "null"]
        userId:
          type: integer
        userName:
          type: string
        createdAt:
          type: string
      required: [id, ingredientId, ingredientName, unit, movementType, quantity, previousStock, newStock, userId, userName, createdAt]
    StockMovementInput:
      type: object
      properties:
        ingredientId:
          type: integer
        movementType:
          type: string
          enum: [in, out, adjustment, opname]
        quantity:
          type: number
        notes:
          type: string
      required: [ingredientId, movementType, quantity]
    RecipeItem:
      type: object
      properties:
        ingredientId:
          type: integer
        ingredientName:
          type: string
        unit:
          type: string
        quantity:
          type: number
        cost:
          type: number
      required: [ingredientId, ingredientName, unit, quantity, cost]
    Recipe:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecipeItem"
        totalHpp:
          type: number
      required: [productId, productName, items, totalHpp]
    RecipeItemInput:
      type: object
      properties:
        ingredientId:
          type: integer
        quantity:
          type: number
      required: [ingredientId, quantity]
    RecipeInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecipeItemInput"
      required: [items]
...