# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREE NINE POS API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface Product {
  id: number;
  name: string;
  code: string;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  description?: string | null;
  sellingPrice: number;
  /** @nullable */
  imageUrl?: string | null;
  status: ProductStatus;
  hpp: number;
  createdAt: string;
  updatedAt: string;
}