# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREE NINE POS API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductUpdateStatus } from './productUpdateStatus';

export interface ProductUpdate {
  name?: string;
  code?: string;
  categoryId?: number;
  description?: string;
  sellingPrice?: number;
  status?: ProductUpdateStatus;
}