# lib/api-zod/src/generated/types/recipeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREE NINE POS API
 * OpenAPI spec version: 0.1.0
 */
import type { RecipeItemInput } from './recipeItemInput';

export interface RecipeInput {
  items: RecipeItemInput[];
}