# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREE NINE POS API
 * OpenAPI spec version: 0.1.0
 */
import type { StockMovementMovementType } from './stockMovementMovementType';

export interface StockMovement {
  id: number;
  ingredientId: number;
  ingredientName: string;
  unit: string;
  movementType: StockMovementMovementType;
  quantity: number;
  previousStock: number;
  newStock: number;
  /** @nullable */
  notes?: string | null;
  userId: number;
  userName: string;
  createdAt: string;
}