# lib/api-zod/src/generated/types/stockMovementInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREE NINE POS API
 * OpenAPI spec version: 0.1.0
 */
import type { StockMovementInputMovementType } from './stockMovementInputMovementType';

export interface StockMovementInput {
  ingredientId: number;
  movementType: StockMovementInputMovementType;
  quantity: number;
  notes?: string;
}